Intent: Unify grammar and graph validation entry points to preserve invariant checks across the API surface Operators involved: [Coherence, Resonance, Coupling] Affected invariants: [#1, #2, #5] Key changes: - Move TNFR graph validators and window guard into tnfr.validation.graph with typing stub and exports in tnfr.validation. - Add compatibility shim in tnfr.utils.validators emitting deprecation warnings and adjust tnfr.utils facade to delegate to the new package. - Update observers, glyph history, runtime, tests, and docs to consume tnfr.validation; register cache metrics logging in the standard observer. Expected risks/dissonances: Import reordering to avoid cycles; mitigated via lazy adapters and regression tests for observers and glyph history. Metrics: No change in baseline C(t), Si, νf; validation suites confirm invariant enforcement. Equivalence map: tnfr.utils.validate_window → tnfr.validation.validate_window; tnfr.utils.run_validators → tnfr.validation.run_validators.
