+++ src/tnfr/utils/__init__.pyi
@@ -79,7 +79,7 @@ from ..io import (
     clear_orjson_param_warnings,
     json_dumps,
 )
-from .validators import run_validators, validate_window
+from ..validation import run_validators, validate_window
 
 __all__ = (
     "IMPORT_LOG",
