+++ src/tnfr/utils/validators.pyi
@@ -1,17 +1,10 @@
 from collections.abc import Mapping
 from typing import Callable, Sequence
 
-from .._compat import TypeAlias
 from ..types import TNFRGraph
-
-ValidatorFunc: TypeAlias = Callable[[TNFRGraph], None]
-NodeData: TypeAlias = Mapping[str, object]
-AliasSequence: TypeAlias = Sequence[str]
+from ..validation.graph import AliasSequence, GRAPH_VALIDATORS, NodeData, ValidatorFunc
 
 __all__: tuple[str, ...]
 
-def __getattr__(name: str) -> object: ...
 def validate_window(window: int, *, positive: bool = ...) -> int: ...
 def run_validators(graph: TNFRGraph) -> None: ...
-
-GRAPH_VALIDATORS: tuple[ValidatorFunc, ...]
